Oprava postfix testů a přidání sin, cos a sqrt.
